# api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost


tags:
  - name: Creates and stores new estate 
    description: APIs for creates and stores new estate in the database. This will receive two integers width and length, representing a palm oil plantation estate. 
  - name: Creates store tree data
    description: APIs for stores tree data in a given estate with the ID, It receives three positive integers x, y, and height
  - name: Get the stats of the tree in the estate with ID
    description: APIs for will simply return the stats of the tree in the estate with ID
  - name: Get Sum distance of the drone monitoring travel
    description: APIs for ill simply return the sum distance of the drone monitoring travel in the estate with ID
  - name: Get Max Distance Drone
    description: APIs for ill simply return the sum distance of the drone monitoring travel in the estate with ID

paths:
  /estate:
    post:
      tags:
        - Creates and stores new estate 
      summary: This Endpoint For Create Estate
      operationId: createEstate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEstateRequest"
              
      responses:
        '201':
          description: Test return
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/CreateEstateResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /estate/tree:
    post:
      tags:
        - Creates store tree data
      summary: APIs for stores tree data in a given estate with the ID, It receives three positive integers x, y, and height
      operationId: createTree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreeRequest"
    
      responses:
        '201':
          description: Test return
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/CreateTreeResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /estate/{Id}stats:
    get:
      tags:
        - Get the stats of the tree in the estate with ID
      summary: APIs for will simply return the stats of the tree in the estate with ID
      operationId: getStats
      parameters:
        - name: Id
          in: path
          required: true
          schema:
            type: string
            format: string
          example: "7e6e5b8e-007a-4bd4-bbcd-57b8e06a029d"

      responses:
        '201':
          description: Test return
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/GetStatsResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /estate/{Id}/drone-plan:
    get:
      tags:
        - Get Sum distance of the drone monitoring travel
      summary:
      operationId: GetDronePlan
      parameters:
        - name: Id
          in: path
          required: true
          schema:
            type: string
            format: string
          example: "7e6e5b8e-007a-4bd4-bbcd-57b8e06a029d"
 
      responses:
        '201':
          description: Test return
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/GetDronePlanResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"



components:
  schemas:
    CreateEstateRequest:
      type: object
      required:
        - width
        - length
      properties:
        width:
          type: integer
        length:
          type: integer
    CreateEstateResponse:
      type: object
      required:
        - id
      properties:
        ID:
          type: string
    CreateTreeRequest:
      type: object
      required:
        - x
        - y 
        - height
        - uuid
      properties:
        x:
          type: integer
        y:
          type: integer
        height:
          type: integer
        uuid:
          type: string
    CreateTreeResponse:
      type: object
      required:
        - id
      properties:
        ID:
          type: string
    GetStatsResponse:
      type: object
      required:
        - count
        - max
        - min
        - median
      properties:
        count:
          type: integer
        max:
          type: integer
        min:
          type: integer
        median:
          type: string
    GetDronePlanResponse:
      type: object
      required:
        - distance
      properties:
        distance:
          type: integer
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string